# Weathervane/backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Weathervane API
  version: "2.0.0"
  description: >
    Production weather forecast backend. Public reads for locations, daily
    forecasts and current conditions; admin writes manage the data.
    All successful responses are wrapped in a `{ "data": ... }` envelope
    (lists add `meta`). Errors use `{ "error": { code, message, details?, requestId } }`.
servers:
  - url: http://localhost:4004
    description: Local (docker compose)
tags:
  - name: Auth
  - name: Locations
  - name: Forecast
  - name: Admin
  - name: Ops
paths:
  /health:
    get:
      tags: [Ops]
      summary: Liveness
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, properties: { status: { type: string, example: ok } } }
  /ready:
    get:
      tags: [Ops]
      summary: Readiness (DB ping)
      responses:
        "200": { description: Ready }
        "503": { description: Database unavailable }
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new USER
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Credentials" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/AuthEnvelope" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login (returns a JWT)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Credentials" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/AuthEnvelope" } } } }
        "401": { $ref: "#/components/responses/Error" }
  /api/auth/me:
    get:
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      summary: Current user
      responses:
        "200": { description: OK }
        "401": { $ref: "#/components/responses/Error" }
  /api/locations:
    get:
      tags: [Locations]
      summary: Search/list locations (paginated)
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: country, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Location" } }
                  meta: { $ref: "#/components/schemas/PageMeta" }
    post:
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      summary: Create a location (ADMIN)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/LocationInput" } } }
      responses:
        "201": { description: Created }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/locations/{id}:
    get:
      tags: [Locations]
      summary: Location detail
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      summary: Update a location (ADMIN)
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { content: { application/json: { schema: { $ref: "#/components/schemas/LocationInput" } } } }
      responses:
        "200": { description: OK }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      summary: Delete a location (ADMIN)
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /api/forecast:
    get:
      tags: [Forecast]
      summary: Daily forecast range for a location
      parameters:
        - { name: locationId, in: query, schema: { type: string, format: uuid } }
        - { name: city, in: query, schema: { type: string } }
        - { name: country, in: query, schema: { type: string } }
        - { name: startDate, in: query, schema: { type: string, format: date } }
        - { name: endDate, in: query, schema: { type: string, format: date } }
      responses:
        "200":
          description: OK (empty range returns an empty array)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/DailyForecast" } }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/forecast/stream:
    get:
      tags: [Forecast]
      summary: SSE stream (not implemented in the REST backend)
      responses:
        "501": { $ref: "#/components/responses/Error" }
  /api/weather/current:
    get:
      tags: [Forecast]
      summary: Current conditions for a location
      parameters:
        - { name: locationId, in: query, schema: { type: string, format: uuid } }
        - { name: city, in: query, schema: { type: string } }
        - { name: country, in: query, schema: { type: string } }
      responses:
        "200": { description: OK }
        "404": { $ref: "#/components/responses/Error" }
  /api/locations/{id}/forecast:
    put:
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      summary: Upsert a daily forecast row (idempotent on location+date)
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/DailyForecastInput" } } }
      responses:
        "200": { description: Upserted }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /api/locations/{id}/current:
    put:
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      summary: Set current weather for a location
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CurrentInput" } } }
      responses:
        "200": { description: OK }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    Credentials:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    AuthEnvelope:
      type: object
      properties:
        data:
          type: object
          properties:
            token: { type: string }
            user: { $ref: "#/components/schemas/User" }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        role: { type: string, enum: [USER, ADMIN] }
    Location:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        city: { type: string }
        country: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        timezone: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    LocationInput:
      type: object
      required: [name, city, country, latitude, longitude, timezone]
      properties:
        name: { type: string }
        city: { type: string }
        country: { type: string }
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        timezone: { type: string, example: "Europe/Istanbul" }
    DailyForecast:
      type: object
      properties:
        date: { type: string, format: date }
        high: { type: number }
        low: { type: number }
        condition: { type: string, enum: [SUNNY, CLOUDY, PARTLY_CLOUDY, RAINY, SNOWY, WINDY, FOGGY, STORMY] }
        precipitationChance: { type: integer, nullable: true }
        humidity: { type: integer, nullable: true }
        windKph: { type: number, nullable: true }
    DailyForecastInput:
      type: object
      required: [date, high, low, condition]
      properties:
        date: { type: string, format: date }
        high: { type: number }
        low: { type: number }
        condition: { type: string, enum: [SUNNY, CLOUDY, PARTLY_CLOUDY, RAINY, SNOWY, WINDY, FOGGY, STORMY] }
        precipitationChance: { type: integer, minimum: 0, maximum: 100 }
        humidity: { type: integer, minimum: 0, maximum: 100 }
        windKph: { type: number, minimum: 0 }
    CurrentInput:
      type: object
      required: [tempC, condition]
      properties:
        tempC: { type: number }
        condition: { type: string, enum: [SUNNY, CLOUDY, PARTLY_CLOUDY, RAINY, SNOWY, WINDY, FOGGY, STORMY] }
        humidity: { type: integer, minimum: 0, maximum: 100 }
        windKph: { type: number, minimum: 0 }
    PageMeta:
      type: object
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: { type: string } }
            requestId: { type: string }