# Weathervane/backend/internal/platform/database/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlc

import (
	"database/sql/driver"
	"fmt"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Condition string

const (
	ConditionSUNNY        Condition = "SUNNY"
	ConditionCLOUDY       Condition = "CLOUDY"
	ConditionPARTLYCLOUDY Condition = "PARTLY_CLOUDY"
	ConditionRAINY        Condition = "RAINY"
	ConditionSNOWY        Condition = "SNOWY"
	ConditionWINDY        Condition = "WINDY"
	ConditionFOGGY        Condition = "FOGGY"
	ConditionSTORMY       Condition = "STORMY"
)

func (e *Condition) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = Condition(s)
	case string:
		*e = Condition(s)
	default:
		return fmt.Errorf("unsupported scan type for Condition: %T", src)
	}
	return nil
}

type NullCondition struct {
	Condition Condition
	Valid     bool // Valid is true if Condition is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullCondition) Scan(value interface{}) error {
	if value == nil {
		ns.Condition, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.Condition.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullCondition) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.Condition), nil
}

type UserRole string

const (
	UserRoleUSER  UserRole = "USER"
	UserRoleADMIN UserRole = "ADMIN"
)

func (e *UserRole) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = UserRole(s)
	case string:
		*e = UserRole(s)
	default:
		return fmt.Errorf("unsupported scan type for UserRole: %T", src)
	}
	return nil
}

type NullUserRole struct {
	UserRole UserRole
	Valid    bool // Valid is true if UserRole is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullUserRole) Scan(value interface{}) error {
	if value == nil {
		ns.UserRole, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.UserRole.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullUserRole) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.UserRole), nil
}

type CurrentWeather struct {
	LocationID uuid.UUID
	TempC      float64
	Condition  Condition
	Humidity   pgtype.Int4
	WindKph    pgtype.Float8
	ObservedAt pgtype.Timestamptz
}

type DailyForecast struct {
	ID                  uuid.UUID
	LocationID          uuid.UUID
	Date                pgtype.Date
	HighC               float64
	LowC                float64
	Condition           Condition
	PrecipitationChance pgtype.Int4
	Humidity            pgtype.Int4
	WindKph             pgtype.Float8
	CreatedAt           pgtype.Timestamptz
	UpdatedAt           pgtype.Timestamptz
}

type Location struct {
	ID        uuid.UUID
	Name      string
	City      string
	Country   string
	Latitude  float64
	Longitude float64
	Timezone  string
	CreatedAt pgtype.Timestamptz
	UpdatedAt pgtype.Timestamptz
}

type User struct {
	ID           uuid.UUID
	Email        string
	PasswordHash string
	Role         UserRole
	CreatedAt    pgtype.Timestamptz
}